update for double

+++ 0x02-functions_nested_loops/10-add.c
@@ -9,7 +9,7 @@
  */
 int add(int a, int b)
 {
-	long sum;
+	double sum;
 
 	sum = a + b;
 	return (sum);
